refactor(evan): add base multi pipelines - multi pipeline rendering in progress
