+++ libraries/evan/headers/evan/RessourceManager.hpp
