+++ libraries/evan/sources/Renderer.cpp
@@ -35,8 +35,12 @@ void evan::Renderer::destroy(VkDevice device)
 {
 	vkDestroyDescriptorPool(device, _descriptorPool, nullptr);
 	vkDestroyDescriptorSetLayout(device, _descriptorSetLayout, nullptr);
-	vkDestroyPipeline(device, _pipeline, nullptr);
-	vkDestroyPipelineLayout(device, _pipelineLayout, nullptr);
+	for (const auto &[_, pipeline]: _pipelines) {
+		vkDestroyPipeline(device, pipeline, nullptr);
+	}
+	for (const auto &[_, pipelineLayout]: _pipelineLayouts) {
+		vkDestroyPipelineLayout(device, pipelineLayout, nullptr);
+	}
 
 	for (Frame &frame: _frames) {
 		frame.destroy(device);
@@ -190,169 +194,147 @@ void evan::Renderer::createDescriptorSetLayout(VkDevice device)
 	}
 }
 
-void evan::Renderer::createGraphicsPipeline(VkDevice device,
+void evan::Renderer::createGraphicsPipelines(VkDevice device,
 											VkRenderPass renderPass,
-											VkSampleCountFlagBits msaaSamples)
+											VkSampleCountFlagBits msaaSamples,
+											const RessourceManager &ressourceManager)
 {
-	auto fragShaderFile = g_assetManager->open(std::string("assets/shaders/frag.spv"));
-	auto vertShaderFile = g_assetManager->open(std::string("assets/shaders/vert.spv"));
+	for (const auto &[id, shader]: ressourceManager.getShaders()) {
+		VkPipelineShaderStageCreateInfo vertShaderStageInfo {};
+		vertShaderStageInfo.sType =
+			VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
+		vertShaderStageInfo.stage  = VK_SHADER_STAGE_VERTEX_BIT;
+		vertShaderStageInfo.module = shader->getVertexShaderModule();
+		vertShaderStageInfo.pName  = "main";
 
-	if (!vertShaderFile || !fragShaderFile) {
-		// If the shaders are not found in the "assets/shaders/" directory, we try
-		// to load them from the "shaders/" directory.
-		fragShaderFile = g_assetManager->open(std::string("shaders/frag.spv"));
-		vertShaderFile = g_assetManager->open(std::string("shaders/vert.spv"));
+		VkPipelineShaderStageCreateInfo fragShaderStageInfo {};
+		fragShaderStageInfo.sType =
+			VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
+		fragShaderStageInfo.stage  = VK_SHADER_STAGE_FRAGMENT_BIT;
+		fragShaderStageInfo.module = shader->getFragmentShaderModule();
+		fragShaderStageInfo.pName  = "main";
+
+		VkPipelineShaderStageCreateInfo shaderStages[] = { vertShaderStageInfo,
+														fragShaderStageInfo };
+
+		VkPipelineVertexInputStateCreateInfo vertexInputInfo {};
+		vertexInputInfo.sType =
+			VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
+
+		auto bindingDescription	   = GPUVertex::getBindingDescription();
+		auto attributeDescriptions = GPUVertex::getAttributeDescriptions();
+
+		vertexInputInfo.vertexBindingDescriptionCount = 1;
+		vertexInputInfo.vertexAttributeDescriptionCount =
+			static_cast<uint32_t>(attributeDescriptions.size());
+		vertexInputInfo.pVertexBindingDescriptions	 = &bindingDescription;
+		vertexInputInfo.pVertexAttributeDescriptions = attributeDescriptions.data();
+
+		VkPipelineInputAssemblyStateCreateInfo inputAssembly {};
+		inputAssembly.sType =
+			VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO;
+		inputAssembly.topology				 = VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
+		inputAssembly.primitiveRestartEnable = VK_FALSE;
+
+		VkPipelineViewportStateCreateInfo viewportState {};
+		viewportState.sType = VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO;
+		viewportState.viewportCount = 1;
+		viewportState.scissorCount	= 1;
+
+		VkPipelineRasterizationStateCreateInfo rasterizer {};
+		rasterizer.sType =
+			VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO;
+		rasterizer.depthClampEnable		   = VK_FALSE;
+		rasterizer.rasterizerDiscardEnable = VK_FALSE;
+		rasterizer.polygonMode			   = VK_POLYGON_MODE_FILL;
+		rasterizer.lineWidth			   = 1.0f;
+		rasterizer.cullMode				   = VK_CULL_MODE_BACK_BIT;
+		rasterizer.frontFace			   = VK_FRONT_FACE_COUNTER_CLOCKWISE;
+		rasterizer.depthBiasEnable		   = VK_FALSE;
+
+		VkPipelineMultisampleStateCreateInfo multisampling {};
+		multisampling.sType =
+			VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO;
+		multisampling.sampleShadingEnable  = VK_FALSE;
+		multisampling.rasterizationSamples = msaaSamples;
+		multisampling.sampleShadingEnable  = VK_TRUE;
+		multisampling.minSampleShading	   = .2f;
+
+		VkPipelineColorBlendAttachmentState colorBlendAttachment {};
+		colorBlendAttachment.colorWriteMask = VK_COLOR_COMPONENT_R_BIT
+			| VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT
+			| VK_COLOR_COMPONENT_A_BIT;
+		colorBlendAttachment.blendEnable = VK_FALSE;
+
+		VkPipelineColorBlendStateCreateInfo colorBlending {};
+		colorBlending.sType =
+			VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
+		colorBlending.logicOpEnable		= VK_FALSE;
+		colorBlending.logicOp			= VK_LOGIC_OP_COPY;
+		colorBlending.attachmentCount	= 1;
+		colorBlending.pAttachments		= &colorBlendAttachment;
+		colorBlending.blendConstants[0] = 0.0f;
+		colorBlending.blendConstants[1] = 0.0f;
+		colorBlending.blendConstants[2] = 0.0f;
+		colorBlending.blendConstants[3] = 0.0f;
+
+		std::vector<VkDynamicState> dynamicStates = { VK_DYNAMIC_STATE_VIEWPORT,
+													VK_DYNAMIC_STATE_SCISSOR,
+													VK_DYNAMIC_STATE_CULL_MODE };
+		VkPipelineDynamicStateCreateInfo dynamicState {};
+		dynamicState.sType = VK_STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO;
+		dynamicState.dynamicStateCount =
+			static_cast<uint32_t>(dynamicStates.size());
+		dynamicState.pDynamicStates = dynamicStates.data();
+
+		VkPipelineLayoutCreateInfo pipelineLayoutInfo {};
+		pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
+		pipelineLayoutInfo.setLayoutCount = 1;
+		pipelineLayoutInfo.pSetLayouts	  = &_descriptorSetLayout;
+
+		if (vkCreatePipelineLayout(device, &pipelineLayoutInfo, nullptr,
+								&_pipelineLayouts[id])
+			!= VK_SUCCESS) {
+			throw std::runtime_error("Failed to create pipeline layout !");
+		}
+
+		VkPipelineDepthStencilStateCreateInfo depthStencil {};
+		depthStencil.sType =
+			VK_STRUCTURE_TYPE_PIPELINE_DEPTH_STENCIL_STATE_CREATE_INFO;
+		depthStencil.depthTestEnable	   = VK_TRUE;
+		depthStencil.depthWriteEnable	   = VK_TRUE;
+		depthStencil.depthCompareOp		   = VK_COMPARE_OP_LESS;
+		depthStencil.depthBoundsTestEnable = VK_FALSE;
+		depthStencil.minDepthBounds		   = 0.0f;
+		depthStencil.maxDepthBounds		   = 1.0f;
+		depthStencil.stencilTestEnable	   = VK_FALSE;
+		depthStencil.front				   = {};
+		depthStencil.back				   = {};
+
+		VkGraphicsPipelineCreateInfo pipelineInfo {};
+		pipelineInfo.sType		= VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
+		pipelineInfo.stageCount = 2;
+		pipelineInfo.pStages	= shaderStages;
+		pipelineInfo.pVertexInputState	 = &vertexInputInfo;
+		pipelineInfo.pInputAssemblyState = &inputAssembly;
+		pipelineInfo.pViewportState		 = &viewportState;
+		pipelineInfo.pRasterizationState = &rasterizer;
+		pipelineInfo.pMultisampleState	 = &multisampling;
+		pipelineInfo.pColorBlendState	 = &colorBlending;
+		pipelineInfo.pDynamicState		 = &dynamicState;
+		pipelineInfo.layout				 = _pipelineLayouts[id];
+		pipelineInfo.renderPass			 = renderPass;
+		pipelineInfo.subpass			 = 0;
+		pipelineInfo.basePipelineHandle	 = VK_NULL_HANDLE;
+		pipelineInfo.pDepthStencilState	 = &depthStencil;
+
+		if (vkCreateGraphicsPipelines(device, VK_NULL_HANDLE, 1, &pipelineInfo,
+									nullptr, &_pipelines[id])
+			!= VK_SUCCESS) {
+			throw std::runtime_error("Failed to create graphics pipeline !");
+		}
+		shader->destroy();
 	}
-
-	std::string vertShaderString = vertShaderFile->content();
-	std::string fragShaderString = fragShaderFile->content();
-
-	std::vector<uint32_t> vertShaderCode(vertShaderString.size()
-										 / sizeof(uint32_t));
-	std::memcpy(vertShaderCode.data(), vertShaderString.data(),
-				vertShaderString.size());
-
-	std::vector<uint32_t> fragShaderCode(fragShaderString.size()
-										 / sizeof(uint32_t));
-	std::memcpy(fragShaderCode.data(), fragShaderString.data(),
-				fragShaderString.size());
-
-	Shader shader(vertShaderCode, fragShaderCode, device);
-
-	VkPipelineShaderStageCreateInfo vertShaderStageInfo {};
-	vertShaderStageInfo.sType =
-		VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
-	vertShaderStageInfo.stage  = VK_SHADER_STAGE_VERTEX_BIT;
-	vertShaderStageInfo.module = shader.getVertexShaderModule();
-	vertShaderStageInfo.pName  = "main";
-
-	VkPipelineShaderStageCreateInfo fragShaderStageInfo {};
-	fragShaderStageInfo.sType =
-		VK_STRUCTURE_TYPE_PIPELINE_SHADER_STAGE_CREATE_INFO;
-	fragShaderStageInfo.stage  = VK_SHADER_STAGE_FRAGMENT_BIT;
-	fragShaderStageInfo.module = shader.getFragmentShaderModule();
-	fragShaderStageInfo.pName  = "main";
-
-	VkPipelineShaderStageCreateInfo shaderStages[] = { vertShaderStageInfo,
-													   fragShaderStageInfo };
-
-	VkPipelineVertexInputStateCreateInfo vertexInputInfo {};
-	vertexInputInfo.sType =
-		VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
-
-	auto bindingDescription	   = Vertex::getBindingDescription();
-	auto attributeDescriptions = Vertex::getAttributeDescriptions();
-
-	vertexInputInfo.vertexBindingDescriptionCount = 1;
-	vertexInputInfo.vertexAttributeDescriptionCount =
-		static_cast<uint32_t>(attributeDescriptions.size());
-	vertexInputInfo.pVertexBindingDescriptions	 = &bindingDescription;
-	vertexInputInfo.pVertexAttributeDescriptions = attributeDescriptions.data();
-
-	VkPipelineInputAssemblyStateCreateInfo inputAssembly {};
-	inputAssembly.sType =
-		VK_STRUCTURE_TYPE_PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO;
-	inputAssembly.topology				 = VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST;
-	inputAssembly.primitiveRestartEnable = VK_FALSE;
-
-	VkPipelineViewportStateCreateInfo viewportState {};
-	viewportState.sType = VK_STRUCTURE_TYPE_PIPELINE_VIEWPORT_STATE_CREATE_INFO;
-	viewportState.viewportCount = 1;
-	viewportState.scissorCount	= 1;
-
-	VkPipelineRasterizationStateCreateInfo rasterizer {};
-	rasterizer.sType =
-		VK_STRUCTURE_TYPE_PIPELINE_RASTERIZATION_STATE_CREATE_INFO;
-	rasterizer.depthClampEnable		   = VK_FALSE;
-	rasterizer.rasterizerDiscardEnable = VK_FALSE;
-	rasterizer.polygonMode			   = VK_POLYGON_MODE_FILL;
-	rasterizer.lineWidth			   = 1.0f;
-	rasterizer.cullMode				   = VK_CULL_MODE_BACK_BIT;
-	rasterizer.frontFace			   = VK_FRONT_FACE_COUNTER_CLOCKWISE;
-	rasterizer.depthBiasEnable		   = VK_FALSE;
-
-	VkPipelineMultisampleStateCreateInfo multisampling {};
-	multisampling.sType =
-		VK_STRUCTURE_TYPE_PIPELINE_MULTISAMPLE_STATE_CREATE_INFO;
-	multisampling.sampleShadingEnable  = VK_FALSE;
-	multisampling.rasterizationSamples = msaaSamples;
-	multisampling.sampleShadingEnable  = VK_TRUE;
-	multisampling.minSampleShading	   = .2f;
-
-	VkPipelineColorBlendAttachmentState colorBlendAttachment {};
-	colorBlendAttachment.colorWriteMask = VK_COLOR_COMPONENT_R_BIT
-		| VK_COLOR_COMPONENT_G_BIT | VK_COLOR_COMPONENT_B_BIT
-		| VK_COLOR_COMPONENT_A_BIT;
-	colorBlendAttachment.blendEnable = VK_FALSE;
-
-	VkPipelineColorBlendStateCreateInfo colorBlending {};
-	colorBlending.sType =
-		VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
-	colorBlending.logicOpEnable		= VK_FALSE;
-	colorBlending.logicOp			= VK_LOGIC_OP_COPY;
-	colorBlending.attachmentCount	= 1;
-	colorBlending.pAttachments		= &colorBlendAttachment;
-	colorBlending.blendConstants[0] = 0.0f;
-	colorBlending.blendConstants[1] = 0.0f;
-	colorBlending.blendConstants[2] = 0.0f;
-	colorBlending.blendConstants[3] = 0.0f;
-
-	std::vector<VkDynamicState> dynamicStates = { VK_DYNAMIC_STATE_VIEWPORT,
-												  VK_DYNAMIC_STATE_SCISSOR,
-												  VK_DYNAMIC_STATE_CULL_MODE };
-	VkPipelineDynamicStateCreateInfo dynamicState {};
-	dynamicState.sType = VK_STRUCTURE_TYPE_PIPELINE_DYNAMIC_STATE_CREATE_INFO;
-	dynamicState.dynamicStateCount =
-		static_cast<uint32_t>(dynamicStates.size());
-	dynamicState.pDynamicStates = dynamicStates.data();
-
-	VkPipelineLayoutCreateInfo pipelineLayoutInfo {};
-	pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-	pipelineLayoutInfo.setLayoutCount = 1;
-	pipelineLayoutInfo.pSetLayouts	  = &_descriptorSetLayout;
-
-	if (vkCreatePipelineLayout(device, &pipelineLayoutInfo, nullptr,
-							   &_pipelineLayout)
-		!= VK_SUCCESS) {
-		throw std::runtime_error("Failed to create pipeline layout !");
-	}
-
-	VkPipelineDepthStencilStateCreateInfo depthStencil {};
-	depthStencil.sType =
-		VK_STRUCTURE_TYPE_PIPELINE_DEPTH_STENCIL_STATE_CREATE_INFO;
-	depthStencil.depthTestEnable	   = VK_TRUE;
-	depthStencil.depthWriteEnable	   = VK_TRUE;
-	depthStencil.depthCompareOp		   = VK_COMPARE_OP_LESS;
-	depthStencil.depthBoundsTestEnable = VK_FALSE;
-	depthStencil.minDepthBounds		   = 0.0f;
-	depthStencil.maxDepthBounds		   = 1.0f;
-	depthStencil.stencilTestEnable	   = VK_FALSE;
-	depthStencil.front				   = {};
-	depthStencil.back				   = {};
-
-	VkGraphicsPipelineCreateInfo pipelineInfo {};
-	pipelineInfo.sType		= VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
-	pipelineInfo.stageCount = 2;
-	pipelineInfo.pStages	= shaderStages;
-	pipelineInfo.pVertexInputState	 = &vertexInputInfo;
-	pipelineInfo.pInputAssemblyState = &inputAssembly;
-	pipelineInfo.pViewportState		 = &viewportState;
-	pipelineInfo.pRasterizationState = &rasterizer;
-	pipelineInfo.pMultisampleState	 = &multisampling;
-	pipelineInfo.pColorBlendState	 = &colorBlending;
-	pipelineInfo.pDynamicState		 = &dynamicState;
-	pipelineInfo.layout				 = _pipelineLayout;
-	pipelineInfo.renderPass			 = renderPass;
-	pipelineInfo.subpass			 = 0;
-	pipelineInfo.basePipelineHandle	 = VK_NULL_HANDLE;
-	pipelineInfo.pDepthStencilState	 = &depthStencil;
-
-	if (vkCreateGraphicsPipelines(device, VK_NULL_HANDLE, 1, &pipelineInfo,
-								  nullptr, &_pipeline)
-		!= VK_SUCCESS) {
-		throw std::runtime_error("Failed to create graphics pipeline !");
-	}
-	shader.destroy();
 }
 
 void evan::Renderer::createDescriptorPool(VkDevice device,
